lib-sieve: util: program-client-local: Changed link between parent and child process from a socketpair to a double pipe.
Turns out /dev/stdin, /dev/stdout, and /dev/fd/XX don't work on Linux if the underlying FD is a socket.
Since some programs may rely on those symbolic links to work, the local client is amended to use two pipes instead.
This will not help for programs invoked in the remote script service, since a socket is required for the connection.