lib-sieve: Added a "sieve_user_email" setting that configures the user's primary email address.
This is used mainly when no other means of obtaining the user's email is available, such as when there is no envelope (in IMAP).
Before, the postmaster address or the null "<>" address was used as a fallback for sending messages.
The redirect and enotify commands can be explicitly configured to use what is configured with sieve_user_email.