LDA Sieve plugin: Implemented support for a "discard script" that is run when the message is going to be discarded.
This adds a new "sieve_discard" setting that points to the location of the "discard script".
Once the normal scripts finish executing and the result is that the message will be discarded without being delivered anywhere, the "discard script" is executed.
The discard script can prevent dropping the message, by executing alternative actions.
If the discard script does nothing, the message is still discarded.