lib-sieve: vacation extension: Allow ignoring the envelope sender while composing the "To:" header for the reply.
This adds a new setting "sieve_vacation_to_header_ignore_envelope".
With this setting enabled, the "To:" header is always composed from the first "Sender", "Resent-From" or "From" header found in the message (in that order).
Normally, the "To:" header is composed from the address found in the "Sender", "Resent-From" or "From" headers that is equal to the envelope sender.
If none is then found, the bare envelope sender is used.
The new setting allows ignoring the envelope, which is useful e.g. when SRS is used.