lib-sieve: imap4flags extension: Fix binary corruption occurring when setflag/addflag/removeflag flag-list is a variable.
The original implementation checked the first operand for being a variable.
Obviously, when the assigned flag string is just a variable, this doesn't work.
This causes the flag list to be recognized erroneously as a variable flags
string. Fixed the problem by explicitly emitting an omitted operand when there
is no second argument for a setflag/addflag/removeflag.

Because the new byte code format is incompatible, the extension version is
bumped to 1.