lib-sieve: util: rfc2822: Fix assert panic occurring in rfc2822_header_append().
Panic was: "Buffer write out of range"

With some rather weird (sender-provided!) input, the header folding algorithm
got confused, causing a pointer to the start of the current line to exceed the
parsing pointer. This caused str_append_data() to be called with a negative
size. Added an assertion to make any future similar problems more obvious.