Make it easier to differentiate which regex failed
It took me quite some time to figure out which regex of my spamtest
configuration failed. And that was more of an accident because I switched
to sieve_spamtest_max_value for a moment and suddenly it started
working. Having the different log messages would have pointed me to the
problem directly.