lib-sieve: redirect action: Implement additional protection against mail loops.
Also check the X-Sieve-Redirected-From header for our own e-mail addresses. This
header is added by the redirect action itself and in a mail loop it would see
that same header with that same content. This is less reliable than the other
mail loop detection (sender may set such a header), so, unlike the existing loop
detection based on the duplicate db, the implicit keep is not canceled when the
new loop detection is triggered.