lib-sieve: plugins: enotify: Allow UTF-8 subject
If the sieve enotify command is invoked with a subject (either in the
notification URL or as the "body" of the enotify command), a URI
validation routine may reject it with

    error: notify action: invalid mailto URI: invalid header field body

This is an unnecessary failure, since the enotify command would
subsequently sanitize illegal Unicode characters and MIME-encode the
Subject: header correctly anyway.

Furthermore, the error message is confusing: it's not the "body"
parameter of the URI that caused the failure, but the value of the
"subject" parameter that caused the rejection.