lib-sieve: sieve-storage - Fix erroneous attempt to read active script for non-personal storage
The sieve_storage_active_script_open() can also be called for a default storage,
but in that case it should only open the script directly using the path or the
name specified for the default storage. Before this fix it tried to read the
active script symlink for the default storage, which makes no sense and causes
incorrect and confusing errors.

Add a few asserts to make sure similar issues are noticed early.